fix: Verify and fix misconceptions for students 41-50 (200140-200149)
Fixed mismatches:
- Walker_Charlotte_200141 Q2: Added VAR001 (precedence error)
- Lewis_Avery_200144 Q3: Added VAR001 (swapped Math.pow args)
- Torres_Henry_200145 Q1: Added INPUT003 (not reading t)
- Torres_Henry_200145 Q3: Added DT003 (nextInt for doubles)
- Carter_Wyatt_200147 Q4: Added OTHER002 (hardcoded values)

+++ authentic_seeded/Smith_Matthew_200149_DT001/Q1.java
@@ -1,16 +1,16 @@
 package assignment2;
+
 import java.util.Scanner;
 
-public class Q1
-{
-    // System.out.println("Debug: " + v0);
+public class Q1 {
+   // System.out.println("Debug: " + v0);
    public static void main(String[] args) {
       Scanner scan = new Scanner(System.in);
       System.out.print("Enter v0, v1, and t: ");
-      double startVelocity=scan.nextDouble();
-      double endVelocity=scan.nextDouble();
-      double time=scan.nextDouble();
-      double acceleration=(endVelocity-startVelocity)/time;
+      double startVelocity = scan.nextDouble();
+      double endVelocity = scan.nextDouble();
+      double time = scan.nextDouble();
+      double acceleration = (int) (endVelocity - startVelocity) / (int) time;
       System.out.println("The average acceleration is " + acceleration);
    }
 }
