+++ authentic_seeded/Torres_Daniel_200148_Mixed/Q2.java
@@ -1,13 +1,14 @@
 import java.util.Scanner;
+
 public class Q2 {
   public static void main(String[] args) {
     Scanner scanner = new Scanner(System.in);
     System.out.print("Enter the driving distance in miles: "); // print output
-    double distance = scanner.nextDouble();
+    double distance = scanner.nextInt();
     System.out.print("Enter miles per gallon: ");
-    double milesPerGallon = scanner.nextDouble();
+    double milesPerGallon = scanner.nextInt();
     System.out.print("Enter price in $ per gallon: ");
-    double price = scanner.nextDouble();
+    double price = scanner.nextInt();
     double cost = (distance / milesPerGallon) * price;
     System.out.println("The cost of driving is $" + cost); // calculate result
   }
