+++ authentic_seeded/Torres_Henry_200145_Mixed/Q3.java
@@ -3,17 +3,16 @@
 // Assignment 2
 
 import java.util.Scanner;
-public class Q3
-{
-  public static void main(String[] args)
-  {
+
+public class Q3 {
+  public static void main(String[] args) {
     Scanner scanner = new Scanner(System.in);
     System.out.print("Enter x1 and y1: ");
-    double x1 = scanner.nextDouble();
-    double y1 = scanner.nextDouble();
+    double x1 = scanner.nextInt();
+    double y1 = scanner.nextInt();
     System.out.print("Enter x2 and y2: ");
-    double x2 = scanner.nextDouble();
-    double y2 = scanner.nextDouble();
+    double x2 = scanner.nextInt();
+    double y2 = scanner.nextInt();
     double distance = Math.sqrt(Math.pow(x2 - x1, 2) + Math.pow(y2 - y1, 2));
     System.out.println("The distance of the two points is " + distance);
   }
